add factory and refactor tests
